Task[#5912] Remove useless type conversion in adt7408.
git-svn-id: svn://localhost/svn/linux-kernel/trunk@8554 526b6c2d-f592-4532-a319-5dd88ccb003d

+++ drivers/staging/iio/adc/adt7408.c
@@ -694,7 +694,7 @@ static inline ssize_t adt7408_set_t_bound(struct device *dev,
 		data = (ADT7408_BOUND_VALUE_SIGN << 1) - data;
 	data <<= ADT7408_BOUND_VALUE_OFFSET;
 
-	ret = adt7408_i2c_write(chip, bound_reg, (u8)data);
+	ret = adt7408_i2c_write(chip, bound_reg, data);
 	if (ret)
 		return -EIO;
